bin/destroy: add support to delete bibtex compile files
I didn't know that bibtex will be use when I made the first
"version" of this script. Now it's possible to delete all the
files that biblatex generate when it compiles. Of course, files
like .bib or .tex will no been deleted, and the directory docs
also will be ignored in order to don't destroy the .pdf documents
inside it.